+++ src/main/java/github/f1angz/tmtb/service/TaskService.java
@@ -1,0 +1,12 @@
+package github.f1angz.tmtb.service;
+
+import github.f1angz.tmtb.entity.Task;
+import github.f1angz.tmtb.entity.TelegramUser;
+
+import java.util.List;
+
+public interface TaskService {
+
+    void addTask(Task task);
+    List<Task> showTasks(String chatId);
+}