+++ RELEASE_NOTES.md
@@ -8,3 +8,9 @@
 ## 0.2.0 - SNAPSHOT
 
 * Implemented Command pattern for handling Telegram Bot commands
+
+## 0.3.0 - SNAPSHOT
+
+* Added repository layer
+* Added a new command /add, with which you can add a task
+* Added a new command /show, with which you can see a list of all added tasks